# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Restaurant QR Table Ordering System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get orders (staff)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, preparing, ready, served, paid, cancelled]
        - name: tableId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order (guest)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (staff)
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/table/{tableToken}:
    get:
      operationId: getTableOrders
      tags: [orders]
      summary: Get orders for a guest table session
      parameters:
        - name: tableToken
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Table orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        menuItemId:
          type: integer
        menuItemName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        notes:
          type: string
      required: [id, menuItemId, menuItemName, quantity, unitPrice]
    Order:
      type: object
      properties:
        id:
          type: integer
        tableId:
          type: integer
        tableName:
          type: string
        tableToken:
          type: string
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, served, paid, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalAmount:
          type: number
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, tableId, tableName, tableToken, status, items, totalAmount, createdAt]
    CreateOrderRequest:
      type: object
      properties:
        tableToken:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              menuItemId:
                type: integer
              quantity:
                type: integer
              notes:
                type: string
            required: [menuItemId, quantity]
        notes:
          type: string
      required: [tableToken, items]
    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [confirmed, preparing, ready, served, paid, cancelled]
      required: [status]
...